requirements.txt
numba==0.61.0
numpy==2.1.3
pyserial==3.5
mss==9.0.2